+++ src/mopidy/audio/__init__.py
@@ -1,12 +1,10 @@
 from mopidy.audio._actor import Audio, AudioProxy
 from mopidy.audio._listener import AudioListener
 from mopidy.audio._utils import supported_uri_schemes
-from mopidy.types import PlaybackState
 
 __all__ = [
     "Audio",
     "AudioListener",
     "AudioProxy",
-    "PlaybackState",  # Re-exported for backwards compatibility.
     "supported_uri_schemes",
 ]